+++ src/exec_command.h
@@ -12,6 +12,6 @@
  * @param argc Number of arguments
  * @param argv Array of arguments
 */
-void execCommand(int *pathc, char pathv[MAX_PATHS][MAX_LINE], int argc, char argv[MAX_ARGS][MAX_LINE]);
+void execCommand(size_t *pathc, char pathv[MAX_PATHS][MAX_LINE], int argc, char argv[MAX_ARGS][MAX_LINE]);
 
 #endif
